cascade support end2end and finetune from faster rcnn

+++ lib/rpn/rpn.py
@@ -197,7 +197,7 @@ def assign_anchor(feat_shape, gt_boxes, im_info, cfg, feat_stride=16,
 
     bbox_targets = np.zeros((len(inds_inside), 4), dtype=np.float32)
     if gt_boxes.size > 0:
-        bbox_targets[:] = bbox_transform(anchors, gt_boxes[argmax_overlaps, :4])
+        bbox_targets[:] = bbox_transform(anchors.astype(np.float32), gt_boxes[argmax_overlaps, :4].astype(np.float32))
 
     bbox_weights = np.zeros((len(inds_inside), 4), dtype=np.float32)
     bbox_weights[labels == 1, :] = np.array(cfg.TRAIN.RPN_BBOX_WEIGHTS)
